Added time widget (fixes #94).
