+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,8 @@
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
-        android:supportsRtl="true">
+        android:supportsRtl="true"
+        android:name=".ForecastieApp">
         <activity
             android:name=".MainActivity"
             android:label="@string/app_name"
@@ -27,13 +28,21 @@
             android:name=".SettingsActivity"
             android:theme="@style/AppTheme"/>
 
-        <receiver android:name="WidgetProvider">
+        <receiver android:name=".widgets.ExtensiveWidgetProvider">
             <intent-filter>
                 <action android:name="android.appwidget.action.APPWIDGET_UPDATE" />
             </intent-filter>
             <meta-data
                 android:name="android.appwidget.provider"
-                android:resource="@xml/weather_widget_info" />
+                android:resource="@xml/extensive_widget" />
+        </receiver>
+        <receiver android:name=".widgets.TimeWidgetProvider">
+            <intent-filter>
+                <action android:name="android.appwidget.action.APPWIDGET_UPDATE" />
+            </intent-filter>
+            <meta-data
+                android:name="android.appwidget.provider"
+                android:resource="@xml/time_widget" />
         </receiver>
         <service
             android:name=".DashClockWeatherExtension"
