+++ app/src/main/res/layout/extensive_widget.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:card_view="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/widgetRoot"
@@ -22,9 +22,9 @@
                 android:id="@+id/widgetTemperature"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="0 °C"
+                tools:text="0 °C"
                 android:textColor="@android:color/white"
-                android:textSize="36dp" />
+                android:textSize="36sp" />
 
             <ImageButton
                 android:id="@+id/widgetButtonRefresh"
@@ -33,25 +33,26 @@
                 android:background="@android:color/transparent"
                 android:src="@drawable/ic_refresh_white_24dp"
                 android:layout_marginTop="14dp"
-                android:layout_marginLeft="16dp"/>
+                android:layout_marginLeft="16dp"
+                android:layout_marginStart="16dp"/>
         </LinearLayout>
 
         <TextView
             android:id="@+id/widgetDescription"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="No data"
+            tools:text="No data"
             android:textColor="@android:color/white"
-            android:textSize="16dp" />
+            android:textSize="16sp" />
 
         <TextView
             android:id="@+id/widgetCity"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="8dp"
-            android:text="City name, AB"
+            tools:text="City name, AB"
             android:textColor="@android:color/white"
-            android:textSize="16dp"
+            android:textSize="16sp"
             android:textStyle="bold" />
 
         <TextView
@@ -59,47 +60,47 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="8dp"
-            android:text="Wind: 0 m/s"
+            tools:text="Wind: 0 m/s"
             android:textColor="@android:color/white"
-            android:textSize="16dp" />
+            android:textSize="16sp" />
 
         <TextView
             android:id="@+id/widgetPressure"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="Pressure: 0 hpa"
+            tools:text="Pressure: 0 hpa"
             android:textColor="@android:color/white"
-            android:textSize="16dp" />
+            android:textSize="16sp" />
 
         <TextView
             android:id="@+id/widgetHumidity"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="Humidity: 0 %"
+            tools:text="Humidity: 0 %"
             android:textColor="@android:color/white"
-            android:textSize="16dp" />
+            android:textSize="16sp" />
 
         <TextView
             android:id="@+id/widgetSunrise"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="Sunrise: 00:00"
+            tools:text="Sunrise: 00:00"
             android:textColor="@android:color/white"
-            android:textSize="16dp" />
+            android:textSize="16sp" />
 
         <TextView
             android:id="@+id/widgetSunset"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="Sunset: 00:00"
+            tools:text="Sunset: 00:00"
             android:textColor="@android:color/white"
-            android:textSize="16dp" />
+            android:textSize="16sp" />
 
         <TextView
             android:id="@+id/widgetLastUpdate"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:textSize="16dp" />
+            android:textSize="16sp" />
     </LinearLayout>
 
     <ImageView
@@ -107,6 +108,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginRight="1dp"
+        android:layout_marginEnd="1dp"
         android:layout_alignParentEnd="true"
         android:layout_alignParentRight="true"
         android:layout_alignParentTop="true" />
