+++ app/src/main/res/xml/extensive_widget.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
+    android:minWidth="270dp"
+    android:minHeight="120dp"
+    android:updatePeriodMillis="1800000"
+    android:initialLayout="@layout/extensive_widget"
+    android:resizeMode="horizontal|vertical"
+    android:widgetCategory="home_screen|keyguard"
+    android:previewImage="@mipmap/ic_launcher">
+</appwidget-provider>
