+++ app/src/main/res/xml/time_widget.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
     android:minWidth="270dp"
-    android:minHeight="120dp"
+    android:minHeight="80dp"
     android:updatePeriodMillis="1800000"
-    android:initialLayout="@layout/weather_widget"
+    android:initialLayout="@layout/time_widget"
     android:resizeMode="horizontal|vertical"
     android:widgetCategory="home_screen|keyguard"
     android:previewImage="@mipmap/ic_launcher">
